requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy>=1.26.0
tensorflow>=2.16.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
cryptography==42.0.0
